+++ R/store.R
@@ -445,15 +445,28 @@ ragnar_store_inspect <- function(store, ...) {
 #' @param host Host to run the Embedding Atlas server on.
 #' @param launch.browser Whether to launch the browser automatically.
 #'
-#' @note This function requires the `embedding-atlas` Python package.
-#' Make sure you have it installed in your reticulate Python environment.
-#' It also uses `arrow` to transfer data from the DuckDB store to Python.
+#' @note This function requires the `embedding-atlas` Python package (>= 0.20.0)
+#' in your reticulate Python environment, the `duckdb` R package (>= 1.4.0),
+#' and `nanoarrow` (>= 0.8.0) to transfer data from the DuckDB store to Python.
 #'
 #' @examples
 #' \dontrun{
-#' # Connect or create a store
-#' store <- ragnar_store_connect(':memory:')
+#' # Start Ollama, then run this in a terminal:
+#' # ollama pull embeddinggemma:300m-qat-q4_0
+#'
+#' # Create an in-memory store using local embeddings
+#' store <- ragnar_store_create(
+#'   embed = embed_ollama(model = "embeddinggemma:300m-qat-q4_0")
+#' )
+#'
+#' # Read and embed a chapter from R for Data Science
+#' chunks <- "https://r4ds.hadley.nz/data-transform.html" |>
+#'   read_as_markdown() |>
+#'   markdown_chunk()
+#' ragnar_store_insert(store, chunks)
+#'
 #' # Launch the Embedding Atlas app
+#' # Interrupt R (Esc or Ctrl+C) to stop the server
 #' ragnar_store_atlas(store)
 #' }
 #'
@@ -466,11 +479,14 @@ ragnar_store_atlas <- function(
   port = 3030,
   launch.browser = interactive()
 ) {
+  rlang::check_installed("nanoarrow", version = "0.8.0")
+
   # duckdb R version need to match the Python version at least in major.minor
   duckdb_version <- numeric_version(utils::packageVersion("duckdb"))
 
   reticulate::py_require(c(
-    "embedding-atlas",
+    "embedding-atlas>=0.20.0",
+    "nanoarrow",
     sprintf("duckdb==%s.%s.*", duckdb_version$major, duckdb_version$minor),
     "numba>=0.62.0rc2",
     "llvmlite>=0.45.0rc2"
@@ -479,11 +495,10 @@ ragnar_store_atlas <- function(
     "_ragnartools.atlas",
     system.file("python", package = "ragnar")
   )
-  df <- duckdb::duckdb_fetch_arrow(DBI::dbSendQuery(
+  df <- DBI::dbGetQueryArrow(
     store@con,
-    "SELECT * FROM chunks;",
-    arrow = TRUE
-  ))
+    "SELECT * FROM chunks;"
+  )
   join_thread <- atlas$run_embedding_atlas(df, host, as.integer(port))
   if (launch.browser) {
     utils::browseURL(sprintf("http://%s:%d", host, port))